@@ -1,4 +1,4 @@
-class CreateBlogAuthors < ActiveRecord::Migration
+class CreateBlogAuthors < ActiveRecord::Migration[5.0]
   def change
     create_table :blog_authors do |t|
       t.string :name
