@@ -1,4 +1,4 @@
-class CreateParents < ActiveRecord::Migration
+class CreateParents < ActiveRecord::Migration[5.0]
   def change
     create_table :parents do |t|
       t.string :name
