@@ -1,4 +1,4 @@
-class CreateChildren < ActiveRecord::Migration
+class CreateChildren < ActiveRecord::Migration[5.0]
   def change
     create_table :children do |t|
       t.string :name
