@@ -1,4 +1,4 @@
-class AddSourceToCompanies < ActiveRecord::Migration
+class AddSourceToCompanies < ActiveRecord::Migration[5.0]
   def change
     add_column :companies, :source, :string, :default => "web"
   end
