@@ -1,21 +1,21 @@
 def create_children
   {
-    :child_one => FactoryGirl.create(:child_one),
-    :child_two => FactoryGirl.create(:child_two),
-    :child_three => FactoryGirl.create(:child_three)
+    :child_one => FactoryBot.create(:child_one),
+    :child_two => FactoryBot.create(:child_two),
+    :child_three => FactoryBot.create(:child_three)
   }
 end
 
 def create_people
   {
-    :person_one => FactoryGirl.create(:person_one),
-    :person_two => FactoryGirl.create(:person_two),
-    :person_three => FactoryGirl.create(:person_three)
+    :person_one => FactoryBot.create(:person_one),
+    :person_two => FactoryBot.create(:person_two),
+    :person_three => FactoryBot.create(:person_three)
   }
 end
 
 def create_blog_authors
   {
-    :author_one => FactoryGirl.create(:author_one)
+    :author_one => FactoryBot.create(:author_one)
   }
 end
