@@ -5,6 +5,7 @@ require 'action_mailer/railtie'
 require 'sprockets/railtie'
 
 begin
+  require "#{CI_ORM}"
   require "#{CI_ORM}/railtie"
 rescue LoadError # rubocop:disable HandleExceptions
 end
@@ -24,8 +25,5 @@ module DummyApp
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
     # config.time_zone = 'Central Time (US & Canada)'
-
-    # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true if Rails.version >= '4.2' && CI_ORM == :active_record
   end
 end
